move activities to separately files
