+++ rwengine/src/ai/CharacterController.hpp
@@ -141,93 +141,4 @@ public:
     }
 };
 
-#define DECL_ACTIVITY(activity_name)                     \
-    static constexpr auto ActivityName = #activity_name; \
-    std::string name() const {                           \
-        return ActivityName;                             \
-    }
-
-// TODO: Refactor this with an ugly macro to reduce code dup.
-
-/**
- * @brief Activities for CharacterController behaviour
- *
- * @todo Move into ControllerActivities.hpp or equivelant
- */
-namespace Activities {
-struct GoTo : public CharacterController::Activity {
-    DECL_ACTIVITY(GoTo)
-
-    glm::vec3 target;
-    bool sprint;
-
-    GoTo(const glm::vec3& target, bool _sprint = false)
-        : target(target), sprint(_sprint) {
-    }
-
-    bool update(CharacterObject* character, CharacterController* controller);
-
-    bool canSkip(CharacterObject*, CharacterController*) const {
-        return true;
-    }
-};
-
-struct Jump : public CharacterController::Activity {
-    DECL_ACTIVITY(Jump)
-
-    bool jumped;
-
-    Jump() : jumped(false) {
-    }
-
-    bool update(CharacterObject* character, CharacterController* controller);
-};
-
-struct EnterVehicle : public CharacterController::Activity {
-    DECL_ACTIVITY(EnterVehicle)
-
-    VehicleObject* vehicle;
-    int seat;
-
-    enum {
-        ANY_SEAT = -1  // Magic number for any seat but the driver's.
-    };
-
-    bool entering;
-
-    EnterVehicle(VehicleObject* vehicle, int seat = 0)
-        : vehicle(vehicle), seat(seat), entering(false) {
-    }
-
-    bool canSkip(CharacterObject* character,
-                 CharacterController*) const override;
-
-    bool update(CharacterObject* character, CharacterController* controller);
-};
-
-struct ExitVehicle : public CharacterController::Activity {
-    DECL_ACTIVITY(ExitVehicle)
-
-    const bool jacked;
-
-    ExitVehicle(bool jacked_ = false) : jacked(jacked_) {
-    }
-
-    bool update(CharacterObject* character, CharacterController* controller);
-};
-
-struct UseItem : public CharacterController::Activity {
-    DECL_ACTIVITY(UseItem)
-
-    int itemslot;
-    bool fired = false;
-    float power = 0.f;
-
-    UseItem(int slot) : itemslot(slot) {
-    }
-
-    bool update(CharacterObject* character, CharacterController* controller);
-};
-}
-
 #endif
