+++ rwengine/src/script/modules/GTA3ModuleImpl.inl
@@ -7,6 +7,9 @@
 #include <engine/GameData.hpp>
 #include <engine/Animator.hpp>
 #include <core/Logger.hpp>
+#include <ai/activity/EnterVehicle.hpp>
+#include <ai/activity/ExitVehicle.hpp>
+#include <ai/activity/GoTo.hpp>
 #include <ai/PlayerController.hpp>
 #include <data/CutsceneData.hpp>
 
@@ -5524,7 +5527,7 @@ void opcode_01d3(const ScriptArguments& args, const ScriptCharacter character, c
 	RW_UNUSED(vehicle);
 	RW_UNUSED(args);
 	character->controller->skipActivity();
-	character->controller->setNextActivity(new Activities::ExitVehicle);
+	character->controller->setNextActivity(new ExitVehicle);
 }
 
 /**
@@ -5537,7 +5540,7 @@ void opcode_01d3(const ScriptArguments& args, const ScriptCharacter character, c
 void opcode_01d4(const ScriptArguments& args, const ScriptCharacter character, const ScriptVehicle vehicle) {
 	RW_UNUSED(args);
 	character->controller->skipActivity();
-	character->controller->setNextActivity(new Activities::EnterVehicle(vehicle,Activities::EnterVehicle::ANY_SEAT));
+	character->controller->setNextActivity(new EnterVehicle(vehicle,EnterVehicle::ANY_SEAT));
 }
 
 /**
@@ -5549,7 +5552,7 @@ void opcode_01d4(const ScriptArguments& args, const ScriptCharacter character, c
 */
 void opcode_01d5(const ScriptArguments& args, const ScriptCharacter character, const ScriptVehicle vehicle) {
 	RW_UNUSED(args);
-	character->controller->setNextActivity(new Activities::EnterVehicle(vehicle));
+	character->controller->setNextActivity(new EnterVehicle(vehicle));
 }
 
 /**
@@ -6329,10 +6332,10 @@ void opcode_0211(const ScriptArguments& args, const ScriptCharacter character, S
 	if( character->isInVehicle() )
 	{
 		// Since we just cleared the Activities, this will become current immediatley.
-		character->controller->setNextActivity(new Activities::ExitVehicle);
+		character->controller->setNextActivity(new ExitVehicle);
 	}
 	
-	character->controller->setNextActivity(new Activities::GoTo(target));
+	character->controller->setNextActivity(new GoTo(target));
 }
 
 /**
@@ -6784,7 +6787,7 @@ void opcode_0237(const ScriptArguments& args, const ScriptGang gangID, const Scr
 */
 void opcode_0239(const ScriptArguments& args, const ScriptCharacter character, ScriptVec2 coord) {
 	auto target = script::getGround(args, glm::vec3(coord, -100.f));
-	character->controller->setNextActivity(new Activities::GoTo(target, true));
+	character->controller->setNextActivity(new GoTo(target, true));
 }
 
 /**
